fix speed display, location log and kml

+++ NaviUtil/LocationManager.h
@@ -14,7 +14,7 @@
 
 @protocol LocationManagerDelegate <NSObject>
 
--(void) locationUpdate:(CLLocationCoordinate2D) location speed:(int) speed distance:(int) distance heading:(double) heading;
+-(void) locationUpdate:(CLLocationCoordinate2D) location speed:(double) speed distance:(int) distance heading:(double) heading;
 -(void) lostLocationUpdate;
 @end
